sensores-RPI-master/requirements.txt
RPi.GPIO
luma.led_matrix
Adafruit_DHT
